Columbia Buddy
Students are excited to be back on campus and want to meet more people within the university network. After months of staying at home, making new friends might not be as easy as it used to be for them.

Columbia Buddy (ColumBuddy) come as an alternative solution/ platform for Columbia students to meet new people and find a buddy to study together.